# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "meuedital",
        "contact": {
            "email": "contato@jaidsondantas.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://back.meuedital.com/api",
            "description": "Server location {{wamp}}"
        }
    ],
    "paths": {
...
        "/candidate": {
            "get": {
                "tags": [
                    "Candidate"
                ],
                "summary": "GET Candidate",
                "operationId": "App\\Http\\Controllers\\CandidateController::find",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Authorization"
                    },
                    {
                        "$ref": "#/components/parameters/order"
                    },
                    {
                        "$ref": "#/components/parameters/limit"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/where"
                    },
                    {
                        "$ref": "#/components/parameters/context"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    },
                    {
                        "$ref": "#/components/parameters/populate"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    },
                    "404": {
                        "description": "Example not found"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Candidate"
                ],
                "summary": "Registro de um novo Candidato",
                "operationId": "store",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Authorization"
                    },
                    {
                        "$ref": "#/components/parameters/populate"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Miami"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Candidate"
                ],
                "summary": "Deletando Candidato",
                "operationId": "store",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Authorization"
                    },
                    {
                        "$ref": "#/components/parameters/populate"
                    },
                    {
                        "$ref": "#/components/parameters/where"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/candidate/1": {
            "get": {
                "tags": [
                    "Candidate"
                ],
                "summary": "GET Candidate",
                "operationId": "App\\Http\\Controllers\\CandidateController::show",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Authorization"
                    },
                    {
                        "$ref": "#/components/parameters/context"
                    },
                    {
                        "$ref": "#/components/parameters/populate"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    },
                    "404": {
                        "description": "Example not found"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Candidate"
                ],
                "summary": "Atualizando Candidato",
                "operationId": "store",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Authorization"
                    },
                    {
                        "$ref": "#/components/parameters/populate"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Miami"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Candidate"
                ],
                "summary": "Deletando Candidato",
                "operationId": "store",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Authorization"
                    },
                    {
                        "$ref": "#/components/parameters/populate"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "components": {
        "parameters": {
            "order": {
                "name": "order",
                "in": "query",
                "description": "<p>Examples:</p> <pre><em>order: {'principal': 'ASC'}      = ORDER BY entity.principal ASC<br>order: {'criadoEm': 'DESC'}      = ORDER BY entity.criadoEm DESC </em></pre>",
                "required": true,
                "schema": {
                    "type": "string",
                    "default": {
                        "created_at": "ASC"
                    }
                }
            },
            "limit": {
                "name": "limit",
                "in": "query",
                "description": "<p>Examples:</p> <pre><em>limit: 10</em></pre>",
                "required": true,
                "schema": {
                    "type": "integer",
                    "default": 10
                }
            },
            "page": {
                "name": "page",
                "in": "query",
                "description": "<p>Examples:</p> <pre><em>offset: 0</em></pre>",
                "required": true,
                "schema": {
                    "type": "integer",
                    "default": 1
                }
            },
            "where": {
                "name": "where",
                "in": "query",
                "description": "examples: <br/> <pre><em>where: {'id': 'eq:1'}                       | WHERE entity.id = 1<br>where: {'id': 'neq:1'}                      | WHERE entity.id != 1<br>where: {'id': 'like:1%}                     | WHERE entity.id like ('1%')<br>where: {'id': 'notLike:1%'}                 | WHERE entity.id not like ('1%')<br>where: {'id': 'gt:1'}                       | WHERE entity.id  1<br>where: {'id': 'gte:1'}                      | WHERE entity.id = 1<br>where: {'id': 'lt:1'}                       | WHERE entity.id &lt; 1<br>where: {'id': 'lte:1'}                      | WHERE entity.id = 1<br>where: {'id': 'in:[1,2,3,4]'}               | WHERE entity.id in (1,2,3,4)<br>where: {'id': 'notIn:[1,2,3,4]'}            | WHERE entity.id not in (1,2,3,4)<br>where: {'id': 'isNull'}                     | WHERE entity.id isNull<br>where: {'id': 'isNotNull'}                  | WHERE entity.id isNotNull<br>where: [{'id': 'eq:1'},{'id': 'eq:2'}]      | WHERE entity.id = 1 or entity.id = 2</em></pre>",
                "schema": {
                    "type": "string",
                    "default": {
                        "and": [],
                        "or": []
                    }
                }
            },
            "context": {
                "name": "context",
                "in": "query",
                "description": "examples: <br/> <pre><em>{'context': '12345678'}</em></pre>",
                "schema": {
                    "type": "string",
                    "default": "{}"
                }
            },
            "per_page": {
                "name": "per_page",
                "in": "query",
                "description": "examples: <br/> <pre><em>5</em></pre>",
                "schema": {
                    "type": "string",
                    "default": "15"
                }
            },
            "populate": {
                "name": "populate",
                "in": "query",
                "description": "examples: <br/> <pre><em>['populateAll']</em></pre>",
                "schema": {
                    "type": "string",
                    "default": [
                        "populateAll"
                    ]
                }
            },
            "Authorization": {
                "name": "Authorization",
                "in": "header",
                "description": "examples: <br/> <pre><em>Bearer </em></pre>",
                "schema": {
                    "type": "string",
                    "default": "Bearer {{api_token}}"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}